mock rolled ros2bnyahaj reading/writing
